// src/app/pages/models/addbranch.ts
export class addbranch {
    transporter_id: string;
    branch_name: string;
    branch_mail:number;
    branch_contact:number;
    branch_address:string;
    branch_lat:string;
    branch_long:string;
    role_id:string;
    city: string;
    state: string;
    contact_person_details: string;
    name: string;
    mobile: string;
    email_id: string;
    message:string;
  }
  
 